DX-1051 - Empty address lines - This test is already written from DX-1033. This error is only thrown when instantiating an Address object with a missing street. https://auctane.atlassian.net/browse/DX-1051
